requirements.txt
matplotlib
numpy==1.22.3
scipy==1.8
scikit-learn==1.0
h5py==3.6.0
pandas
pyparsing
dataclasses
seaborn==0.11.2
pyarrow
simpy
protobuf==3.20.1
keras==2.8.0
tensorflow==2.8.0
tensorflow-probability==0.16.0
tensorflow-datasets==4.6.0
tensorflow-addons==0.16.1
tensorflow_io==0.25.0
git+https://github.com/samiemostafavi/qsimpy.git
git+https://github.com/samiemostafavi/pr3d.git





